chore: move tests to test dir. cleanup test config format.
