+++ test/upchunk.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from '@open-wc/testing';
 import xhrMock, { sequence, delay } from 'xhr-mock';
 import xhr from 'xhr';
-import { UpChunk, createUpload, UpChunkOptions } from './upchunk';
+import { UpChunk, createUpload, UpChunkOptions } from '../src/upchunk';
 
 describe('integration', () => {
   const endpoint = `https://this-is-a-fake-url.com/upload/endpoint`;
